// src/app/models/user/participant/report/reportMotive.ts
export enum ReportMotive {
  INAPPROPRIATE_BEHAVIOR = 'INAPPROPRIATE_BEHAVIOR',
  INAPPROPRIATE_CONTENT = 'INAPPROPRIATE_CONTENT',
  OFFENSIVE_LANGUAGE = 'OFFENSIVE_LANGUAGE',
  IMPERSONATION = 'IMPERSONATION',
  ABSENTEEISM = 'ABSENTEEISM',
  HARASSMENT = 'HARASSMENT',
  SPAMMING = 'SPAMMING',
  FRAUD = 'FRAUD',
  OTHER = 'OTHER'
}